Important bugfix for try_update_term
Previously, this function didn't update the leader nor state when
receiving an AE from the node's current round. This had two problems:

- If we are a leaderless follower(e.g, because we observed the current
  term from a vote request/response or an AE response), this function
  wouldn't update our leader. It turns out this functionality was
  implemented at `FollowerState::handle_append_entries`, but I moved
  it here for consistency

- If we are a candidate and someone else won the election in the current
  term, we could remain in candidate state and try another election,
  this behaviour might repeat causing an election loop.

`try_update_term` was expanded to deal with these, and be more
informative. It also checks for election safety property, instead of
`FollowerState::handle_append_entries`, for sake of consistency