Reset election timer upon getting RV with higher term
Raft spec dictates that upon seeing an RPC with a higher term, we
should "become a follower".
If we're already a follower there won't really be a state change
in our implementation, so we'll just update the timer.

I'm not entirely sure if this stems from the Raft spec, but I've seen
several implementations that do this